// lexiconTypes/types/fm/teal/alpha/actor/searchActors.ts
/**
 * GENERATED CODE - DO NOT MODIFY
 */
import { HeadersMap, XRPCError } from '@atproto/xrpc'
import { type ValidationResult, BlobRef } from '@atproto/lexicon'
import { CID } from 'multiformats/cid'
import { validate as _validate } from '../../../../../lexicons'
import {
  type $Typed,
  is$typed as _is$typed,
  type OmitKey,
} from '../../../../../util'
import type * as FmTealAlphaActorDefs from './defs.js'

const is$typed = _is$typed,
  validate = _validate
const id = 'fm.teal.alpha.actor.searchActors'

export type QueryParams = {
  /** The search query */
  q: string
  /** The maximum number of actors to return */
  limit?: number
  /** Cursor for pagination */
  cursor?: string
}
export type InputSchema = undefined

export interface OutputSchema {
  actors: FmTealAlphaActorDefs.MiniProfileView[]
  /** Cursor for pagination */
  cursor?: string
}

export interface CallOptions {
  signal?: AbortSignal
  headers?: HeadersMap
}

export interface Response {
  success: boolean
  headers: HeadersMap
  data: OutputSchema
}

export function toKnownErr(e: any) {
  return e
}